Move ListManager proxy code to gc-lists plugin
